updated user specific routes to check for session data. and added notes functionality

+++ flask_app/flask06.py
@@ -38,9 +38,14 @@ with app.app_context():
 @app.route('/')
 @app.route('/index')
 def index():
+    #check if a user is saved in session
+    if session.get('user'):
+        return render_template("index.html", user=session['user'])
+    return render_template("index.html")
+
     #get user from database
-    a_user =  db.session.query(User).filter_by(email='mdiaz11@uncc.edu').one()
-    return render_template('index.html', user=a_user)
+    #a_user =  db.session.query(User).filter_by(email='mdiaz11@uncc.edu').one()
+    #return render_template('index.html', user=a_user)
 
 
 
@@ -60,41 +65,45 @@ def get_notes():
 
 @app.route('/notes/<note_id>')
 def get_note(note_id):
-    #retrieve user from database
-    a_user = db.session.query(User).filter_by(email='mdiaz11@uncc.edu').one()
-    #retrieve note from database
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
-    
-    
-    #a_user = {'name': 'Michael Diaz', 'email':'mogli@uncc.edu'} #Mock
 
-    return render_template('note.html', note=my_note, user=a_user)
+    if session.get('user'):
+
+        #retrieve not from database
+        my_note = db.session.query(Note).filter_by(id=note_id,user_id=session['user_id']).one()
+        
+        return render_template('note.html', note=my_note, use=session['user'])
+    else: return redirect(url_for('login'))
 
 
 
 @app.route('/notes/new', methods = ['GET','POST'])
 def new_note():
-    #check method used for request
-    if request.method == 'POST':
-        #get title data
-        title = request.form['title']
-        #get note data
-        text = request.form['noteText']
+    if session.get('user'):
+        #check method used for request
+        if request.method == 'POST':
+            #get title data
+            title = request.form['title']
+            #get note data
+            text = request.form['noteText']
 
-        #create date stamp
-        from datetime import  date
-        today = date.today()
-        #format date mm/dd/yyyy
-        today = today.strftime("%m-%d-%Y")
+            #create date stamp
+            from datetime import  date
+            today = date.today()
+            #format date mm/dd/yyyy
+            today = today.strftime("%m-%d-%Y")
 
-        newRecord = Note(title, text, today)
-        db.session.add(newRecord)
-        db.session.commit()
+            new_record = Note(title, text, today, session['user_id'])
+            db.session.add(new_record)
+            db.session.commit()
 
-        return redirect(url_for('get_notes'))
+            return redirect(url_for('get_notes'))
+        else:
+            #GET request - show new note from
+            return render_template ('new.html', user=session['user'])
+
     else:
-        a_user = db.session.query(User).filter_by(email='mdiaz11@uncc.edu').one()
-        return render_template ('new.html', user=a_user)
+        #user is not in session redirect to login
+        return redirect(url_for('login'))
 
 
 
@@ -103,44 +112,56 @@ def new_note():
 
 @app.route('/notes/edit/<note_id>', methods = ['GET', 'POST'])
 def update_note(note_id):
-    #check method used for request
-    if request.method == 'POST':
-        #get title data
-        title = request.form['title']
-        #get note data
-        text = request.form['noteText']
-        note = db.session.query(Note).filter_by(id = note_id).one()
-        #update note data
-        note.title = title
-        note.text = text
-        #update note in DB
-        db.session.add(note)
-        db.session.commit()
+    #check if a user is saved in session
+    if session.get('user'):
+        #check method used for request
+        if request.method == 'POST':
+            #get title data
+            title = request.form['title']
+            #get note data
+            text = request.form['noteText']
+            note = db.session.query(Note).filter_by(id = note_id).one()
+            #update note data
+            note.title = title
+            note.text = text
+            #update note in DB
+            db.session.add(note)
+            db.session.commit()
 
 
-        return redirect (url_for('get_notes'))
+            return redirect (url_for('get_notes'))
 
+        else:
+            #GET request - show new note form to edit note
+
+            #retrieve user from database
+            a_user = db.session.query(User).filter_by(email = 'mdiaz11@uncc.edu').one()
+
+            #retrieve note from database
+            my_note = db.session.query(Note).filter_by(id=note_id).one()
+
+            return render_template('new.html', note = my_note, user=session['user'])
     else:
-        #GET request - show new note form to edit note
-        #retrieve user from database
-        a_user = db.session.query(User).filter_by(email = 'mdiaz11@uncc.edu').one()
-
-        #retrieve note from database
-        my_note = db.session.query(Note).filter_by(id=note_id).one()
-
-        return render_template('new.html', note = my_note, user=a_user)
+        #user is not in session redirect to login
+        return redirect(url_for9('login'))
 
 
 
 
 @app.route ('/notes/delete/<note_id>', methods = ['POST'])
 def delete_note(note_id):
-    #retrieve note from database
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
-    db.session.delete(my_note)
-    db.session.commit()
+    #check if a user is saved in session
+    if session.get('user'):
 
-    return redirect (url_for('get_notes'))
+        #retrieve note from database
+        my_note = db.session.query(Note).filter_by(id=note_id).one()
+        db.session.delete(my_note)
+        db.session.commit()
+
+        return redirect (url_for('get_notes'))
+    else:
+        #user is not in session redirect to login
+        return redirect(url_for('login'))
